requirements.txt
Flask==1.1.2
google-cloud-logging==1.15.1
gunicorn==20.0.4